Refactored file and package structure of a project, added annotations to LocalDate fields which are specifies date format
