+++ src/main/java/com/softserve/skillscope/mapper/TalentMapper.java
@@ -1,10 +1,9 @@
 package com.softserve.skillscope.mapper;
 
-import com.softserve.skillscope.talent.model.dto.GeneralTalent;
+import com.softserve.skillscope.talent.model.dto.responce.GeneralTalent;
 import com.softserve.skillscope.talent.model.entity.Talent;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
-import org.springframework.security.core.userdetails.User;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface TalentMapper {
