+++ src/main/java/com/softserve/skillscope/talent/controller/TalentController.java
@@ -1,6 +1,6 @@
 package com.softserve.skillscope.talent.controller;
 
-import com.softserve.skillscope.talent.service.interfaces.TalentService;
+import com.softserve.skillscope.talent.service.TalentService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
