+++ src/main/java/com/softserve/skillscope/talent/service/TalentServiceImpl.java
@@ -1,12 +1,12 @@
-package com.softserve.skillscope.talent.service.impl;
+package com.softserve.skillscope.talent.service;
 
 import com.softserve.skillscope.exception.generalException.BadRequestException;
 import com.softserve.skillscope.mapper.TalentMapper;
 import com.softserve.skillscope.talent.TalentRepository;
-import com.softserve.skillscope.talent.service.interfaces.TalentService;
-import com.softserve.skillscope.talent.model.dto.GeneralTalent;
+import com.softserve.skillscope.talent.model.dto.responce.GeneralTalent;
 import com.softserve.skillscope.talent.model.entity.Talent;
 import com.softserve.skillscope.talent.model.entity.TalentProperties;
+import com.softserve.skillscope.talent.service.TalentService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
