+++ src/main/java/com/softserve/skillscope/talentInfo/model/entity/TalentInfo.java
@@ -1,5 +1,6 @@
 package com.softserve.skillscope.talentInfo.model.entity;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.softserve.skillscope.talent.model.entity.Talent;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotEmpty;
@@ -41,9 +42,10 @@ public class TalentInfo {
 
     @Size(max = 16)
     private String phone;
-//TODO: @SEM resolves problems with age
-//    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
-//    @DateTimeFormat(pattern = "dd.MM.yyyy")
+
+    //TODO: @SEM resolves problems with age
+    //@Illia Kyselov in addition to @S
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
     private LocalDate age;
 
     @Size(max = 64)
